+++ packages/core/src/editing/editing-service.test.ts
@@ -92,6 +92,7 @@ describe('EditingService', () => {
         headers: {
           sc_layoutKind: 'final',
           sc_editMode: 'true',
+          sc_previewMode: 'false',
         },
       }
     );
@@ -141,6 +142,7 @@ describe('EditingService', () => {
         headers: {
           sc_layoutKind: 'final',
           sc_editMode: 'false',
+          sc_previewMode: 'true',
         },
       }
     );
@@ -194,6 +196,7 @@ describe('EditingService', () => {
         headers: {
           sc_layoutKind: 'final',
           sc_editMode: 'true',
+          sc_previewMode: 'false',
         },
       }
     );
@@ -247,6 +250,7 @@ describe('EditingService', () => {
         headers: {
           sc_layoutKind: 'final',
           sc_editMode: 'true',
+          sc_previewMode: 'false',
         },
       }
     );
@@ -292,6 +296,7 @@ describe('EditingService', () => {
         headers: {
           sc_layoutKind: 'shared',
           sc_editMode: 'true',
+          sc_previewMode: 'false',
         },
       }
     );
@@ -399,12 +404,18 @@ describe('EditingService', () => {
     await service.fetchEditingData(editingOptions, fetchOptions);
 
     expect(requestMock.calledOnce).to.be.true;
-    expect(requestMock.firstCall.args[2]).to.deep.equal({
-      ...fetchOptions,
-      headers: {
-        sc_editMode: 'true',
-        sc_layoutKind: LayoutKind.Final,
-      },
+
+    const requestOptions = requestMock.firstCall.args[2];
+
+    expect(requestOptions.retries).to.equal(fetchOptions.retries);
+    expect(requestOptions.fetch).to.equal(fetchOptions.fetch);
+    expect(requestOptions.retryStrategy).to.equal(fetchOptions.retryStrategy);
+    expect(requestOptions.headers).to.deep.equal({
+      Authorization: 'Bearer test-token',
+      'Content-Type': 'application/json',
+      sc_editMode: 'true',
+      sc_layoutKind: LayoutKind.Final,
+      sc_previewMode: 'false',
     });
   });
 });